Refactor App and WebApp to improve modularity by adding AppClassicKomet and AppMenu getters, and streamline menu creation in AppMenu.
