+++ application/src/main/java/dev/ikm/komet/app/AppMenu.java
@@ -1,5 +1,6 @@
 package dev.ikm.komet.app;
 
+import dev.ikm.komet.app.aboutdialog.AboutDialog;
 import javafx.application.Platform;
 import javafx.scene.control.Menu;
 import javafx.scene.control.MenuBar;
@@ -10,7 +11,12 @@ import javafx.scene.input.KeyCombination;
 import javafx.scene.layout.BorderPane;
 import javafx.stage.Stage;
 
+import java.io.IOException;
+import java.util.prefs.BackingStoreException;
+
 import static dev.ikm.komet.app.WebApp.IS_BROWSER;
+import static dev.ikm.komet.kview.mvvm.view.changeset.exchange.GitTask.OperationMode.PULL;
+import static dev.ikm.komet.kview.mvvm.view.changeset.exchange.GitTask.OperationMode.SYNC;
 
 public class AppMenu {
 
@@ -25,7 +31,7 @@ public class AppMenu {
         Menu fileMenu = new Menu("File");
 
         MenuItem about = new MenuItem("About");
-        about.setOnAction(_ -> app.showAboutDialog());
+        about.setOnAction(_ -> showAboutDialog());
         fileMenu.getItems().add(about);
 
         // Importing data
@@ -68,5 +74,78 @@ public class AppMenu {
         menuBar.getMenus().add(windowMenu);
         //hBox.getChildren().add(menuBar);
         Platform.runLater(() -> kometRoot.setTop(menuBar));
+    }
+
+    Menu createExchangeMenu() {
+        Menu exchangeMenu = new Menu("Exchange");
+
+        MenuItem infoMenuItem = new MenuItem("Info");
+        infoMenuItem.setOnAction(actionEvent -> app.getAppGithub().infoAction());
+        MenuItem pullMenuItem = new MenuItem("Pull");
+        pullMenuItem.setOnAction(actionEvent -> app.getAppGithub().executeGitTask(PULL));
+        MenuItem pushMenuItem = new MenuItem("Sync");
+        pushMenuItem.setOnAction(actionEvent -> app.getAppGithub().executeGitTask(SYNC));
+
+        exchangeMenu.getItems().addAll(infoMenuItem, pullMenuItem, pushMenuItem);
+        return exchangeMenu;
+    }
+
+    public void showAboutDialog() {
+        AboutDialog aboutDialog = new AboutDialog();
+        aboutDialog.showAndWait();
+    }
+
+    public void createMenuOptions(BorderPane landingPageRoot) {
+        MenuBar menuBar = new MenuBar();
+
+        Menu fileMenu = new Menu("File");
+        MenuItem about = new MenuItem("About");
+        about.setOnAction(_ -> showAboutDialog());
+        fileMenu.getItems().add(about);
+
+        MenuItem menuItemQuit = new MenuItem("Quit");
+        KeyCombination quitKeyCombo = new KeyCodeCombination(KeyCode.Q, KeyCombination.CONTROL_DOWN);
+        menuItemQuit.setOnAction(actionEvent -> app.quit());
+        menuItemQuit.setAccelerator(quitKeyCombo);
+        fileMenu.getItems().add(menuItemQuit);
+
+        Menu viewMenu = new Menu("View");
+        MenuItem classicKometMenuItem = createClassicKometMenuItem();
+        viewMenu.getItems().add(classicKometMenuItem);
+
+        Menu windowMenu = new Menu("Window");
+        MenuItem minimizeWindow = new MenuItem("Minimize");
+        KeyCombination minimizeKeyCombo = new KeyCodeCombination(KeyCode.M, KeyCombination.CONTROL_DOWN);
+        minimizeWindow.setOnAction(event -> {
+            Stage obj = (Stage) landingPageRoot.getScene().getWindow();
+            obj.setIconified(true);
+        });
+        minimizeWindow.setAccelerator(minimizeKeyCombo);
+        minimizeWindow.setDisable(IS_BROWSER);
+        windowMenu.getItems().add(minimizeWindow);
+
+        Menu exchangeMenu = createExchangeMenu();
+
+        menuBar.getMenus().add(fileMenu);
+        menuBar.getMenus().add(viewMenu);
+        menuBar.getMenus().add(windowMenu);
+        menuBar.getMenus().add(exchangeMenu);
+        landingPageRoot.setTop(menuBar);
+    }
+
+    private MenuItem createClassicKometMenuItem() {
+        MenuItem classicKometMenuItem = new MenuItem("Classic Komet");
+        KeyCombination classicKometKeyCombo = new KeyCodeCombination(KeyCode.K, KeyCombination.CONTROL_DOWN);
+        classicKometMenuItem.setOnAction(actionEvent -> {
+            try {
+                app.getAppClassicKomet().launchClassicKomet();
+            } catch (IOException e) {
+                throw new RuntimeException(e);
+            } catch (BackingStoreException e) {
+                throw new RuntimeException(e);
+            }
+        });
+        classicKometMenuItem.setAccelerator(classicKometKeyCombo);
+        return classicKometMenuItem;
     }
 }
